refactor: moves public track listing from table to list
The track row was rendered as tables but it never really made sense for any listener facing
playlists, and regularly caused issues.

This switches PublicTrackGroupListing and TrackRow to a list flex layout and removes the quick
defensive fix I pushed earlier which was strickly to avoid an issue with the new data table look.

I also added a hover underline on artist names in the track row (much simpler to add now), so users
hovering the artist link see it is a distinct click target before clicking.
