+++ client/src/components/TrackGroup/TrackGroup.tsx
@@ -6,7 +6,7 @@ import ImageWithPlaceholder from "components/common/ImageWithPlaceholder";
 import MarkdownContent from "components/common/MarkdownContent";
 import { MetaCard } from "components/common/MetaCard";
 import SupportArtistPopUp from "components/common/SupportArtistPopUp";
-import PublicTrackGroupListing from "components/common/TrackTable/PublicTrackGroupListing";
+import PublicTrackGroupListing from "components/common/TrackList/PublicTrackGroupListing";
 import WidthContainer from "components/common/WidthContainer";
 import { queryArtist, queryTrackGroup } from "queries";
 import { useTranslation } from "react-i18next";
