+++ client/src/components/Widget/LabelWidget.tsx
@@ -1,6 +1,6 @@
 import ImageWithPlaceholder from "components/common/ImageWithPlaceholder";
 import ScrollFadeOverlay from "components/common/ScrollFadeOverlay";
-import PublicTrackGroupListing from "components/common/TrackTable/PublicTrackGroupListing";
+import PublicTrackGroupListing from "components/common/TrackList/PublicTrackGroupListing";
 import { AudioWrapper } from "components/Player/AudioWrapper";
 import useCurrentTrackHook from "components/Player/useCurrentTrackHook";
 import React from "react";
