+++ client/src/components/Widget/TrackGroupWidget/variants/Strip.tsx
@@ -2,7 +2,7 @@ import ImageWithPlaceholder from "components/common/ImageWithPlaceholder";
 import NextButton from "components/common/NextButton";
 import PrevButton from "components/common/PrevButton";
 import ScrollFadeOverlay from "components/common/ScrollFadeOverlay";
-import PublicTrackGroupListing from "components/common/TrackTable/PublicTrackGroupListing";
+import PublicTrackGroupListing from "components/common/TrackList/PublicTrackGroupListing";
 import React from "react";
 import { useTranslation } from "react-i18next";
